Update index.js
added a hawku const, for the hawku.js file that contains some methods to pull the tablet area, and other details. 

added a new "area" and "areadetails" command. commented out the old area command because it can be used, for example, if someone doesn't want use hawku.  
these are dependent on the 'areaPath' being populated in the config.js file.